+++ Client-Admin/src/gui/app/menu/MenuComponent.fxml
@@ -26,7 +26,7 @@
          <content>
             <AnchorPane>
                <children>
-                  <fx:include source="execution/history/ExecutionComponent.fxml" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0" />
+                  <fx:include source="execution/ExecutionComponent.fxml" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0" />
                </children>
             </AnchorPane>
          </content></Tab>
