+++ Client-Admin/src/gui/app/menu/execution/details/data/PopulationData.java
@@ -1,4 +1,4 @@
-package gui.result.models;
+package gui.app.menu.execution.details.data;
 
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
