+++ Client-Admin/src/gui/app/menu/execution/result/ResultTabComponent.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<VBox xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.app.menu.execution.history.result.ResultTabComponentController">
+<VBox xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.app.menu.execution.result.ResultTabComponentController">
    <children>
       <Label fx:id="executionResultLabel" styleClass="Title" text="Excecution Result" VBox.vgrow="NEVER">
          <font>
