+++ Client-Admin/src/gui/app/menu/execution/result/chart/ChartComponent.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.chart.LineChart?>
 <?import javafx.scene.chart.NumberAxis?>
 
-<LineChart fx:id="chart" prefHeight="250.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.app.menu.execution.history.result.chart.ChartComponentController">
+<LineChart fx:id="chart" prefHeight="250.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.app.menu.execution.result.chart.ChartComponentController">
   <xAxis>
     <NumberAxis fx:id="ticksAxis" label="Ticks" side="BOTTOM" />
   </xAxis>
